Django_Project/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mssql-django==1.1.2
packaging==23.2
pyodbc==5.0.1
pytz==2023.3.post1
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
